+++ src/environment.rs
@@ -66,6 +66,16 @@ pub(crate) fn set_bool<K: AsRef<OsStr>>(key: K, value: bool) {
     set(key, &value_str);
 }
 
+pub(crate) fn set_optional<K: AsRef<OsStr>, V: AsRef<OsStr>>(key: K, value: &Option<V>) -> bool {
+    match value {
+        Some(ref value_ref) => {
+            set(key, value_ref);
+            true
+        }
+        None => false,
+    }
+}
+
 pub(crate) fn get_set<K: AsRef<OsStr>, V: AsRef<OsStr>>(key: K, value: V) -> Option<String> {
     let pre_value = if exists(&key) {
         Some(get_or(&key, ""))