+++ src/environment_test.rs
@@ -191,6 +191,24 @@ fn set_bool_true() {
 }
 
 #[test]
+fn set_optional_none() {
+    let output = set_optional::<&str, &str>("TEST_SET_OPTIONAL_NONE", &None);
+
+    assert!(!output);
+}
+
+#[test]
+fn set_optional_some() {
+    let output = set_optional("TEST_SET_OPTIONAL_SOME", &Some("OPTIONAL VALUE"));
+
+    assert!(output);
+    assert_eq!(
+        env::var("TEST_SET_OPTIONAL_SOME").unwrap(),
+        "OPTIONAL VALUE".to_string()
+    );
+}
+
+#[test]
 fn get_set_exists() {
     env::set_var("TEST_GET_SET_EXISTS", "OLD");
     let output = get_set("TEST_GET_SET_EXISTS", "NEW");